Split EnclaveService into DomainEnclaveService and EnclaveService
